keep rent payer as kora and tidy up the files
